Incremental Model Version 2
Over the course of the project this model changed. Phase 2 and phase 3 were essentially swapped. This required making winner select before the betting formulas because it was easier and could then be regression tested later. It also allowed the game to be structured with the idea that betting would later need to be implemented. Other changes include how player level was calculated and how player instantiation was handled in designing and testing.