Casino UML
I've added the UML for the main simulation function which encapsulates each game. It focuses on the simulation as the central control point of each functionality. the games function as subtasks of the program and the player simulator activates at each round of games through the simulator which passes off the information.