Added Input Error Catch
Beta test comments said that the menu had no error correction when selecting settings for games. I also removed the replay simulation feature because comments said it didn't allow you to re=pick settings and run which caused user grief. If we wanted to change this we would have to change how the casino is instantiated, since the init function uses simMenu().